refactor: home header 폴더 추가 및 파일 위치 변경
